307-subseteq.c: rework int i to size_t i and removed (int)

+++ src/307-subseteq.c
@@ -7,7 +7,7 @@
 #include <stdio.h>
 
 bool Cmd_subseteq(set_t *set1, set_t *set2) {
-    for (int i = 0; i < (int)set2->number_of_elements; ++i) {
+    for (size_t i = 0; i < set2->number_of_elements; ++i) {
         if (!is_element_in_set(set2->elements[i], set1))
             return false;
     }
